refactor code and add test folders and files
